requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.2
django-unfold==0.55.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.8
pillow==11.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
